requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coverage==4.5.4
cytoolz==0.10.0
Django==2.1.5
django-el-pagination==3.2.4
django-mathfilters==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.0
eth-typing==2.1.0
eth-utils==1.6.1
feedparser==5.2.1
iconsdk==1.1.0
idna==2.7
importlib-metadata==0.20
more-itertools==7.2.0
multipledispatch==0.5.0
oauthlib==3.1.0
packaging==19.1
Pillow==5.4.1
pluggy==0.12.0
pprintpp==0.4.0
psycopg2-binary==2.7.6.1
py==1.8.0
pycountry==19.8.18
pycparser==2.19
pycryptodome==3.8.2
pyparsing==2.4.2
PySocks==1.7.1
pytest==5.1.2
pytest-cov==2.7.1
pytest-mock==1.10.4
pytz==2018.9
repoze.lru==0.7
requests==2.20.0
requests-cache==0.4.13
requests-oauthlib==1.2.0
secp256k1==0.13.2
six==1.12.0
soupsieve==1.9.4
toolz==0.10.0
tweepy==3.8.0
urllib3==1.24.3
wcwidth==0.1.7
zipp==0.6.0
